Merging datasets . R
In this script, we merge the individual-level variables from the Eurobarometer (2020–2023) with the aggregated variables from other databases (Eurostat, World Bank, The Economist Intelligence Unit, etc.). This results in a big dataset with more than 100,000 observations and around 70–80 variables. However, this is not the final dataset we will use, as we still need to preprocess it (clean NAs, impute, select, create variables, etc.).