Imputed data sample
Given GitHub's storage limitations, a random sample of the imputed dataset with around 40,000 observations —rather than the full 100,000— is included for practical use. This is the sample of the final dataset. The completed imputed dataset can be found at the URL provided in the .txt document located in this same folder "Preprocess data"